tests/wpt/web-platform-tests/IndexedDB/idbdatabase_transaction.htm

<!DOCTYPE html>
<title>IDBDatabase.transaction() - attempt to open a transaction with invalid scope</title>
<link rel="author" title="Microsoft" href="http://www.microsoft.com">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="support.js"></script>

<script>

var db,
  t = async_test(),
  open_rq = createdb(t);

open_rq.onupgradeneeded = function() {};
open_rq.onsuccess = function(e) {
    db = e.target.result;

    assert_throws('NotFoundError', function() { db.transaction('non-existing'); });
    t.done();
};

</script>

<div id="log"></div>
